Separation of jvm-specific code from TableQueryUpdateConnectionTest
